Added cache system
Now it adds same svg icon elements into a loading queue if the svg wasn't previous loaded.
Once the svg is loaded it's cached and applies the configuration to every element related to this svg. prints into the html and removes if from the cache.
If one icon want's to load and its svg is stored at cache it just applies it.